refactor: remove redis
